Move searchBar to containers
