+++ src/__test__/components/SearchBar.test.js
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
 import { create } from 'react-test-renderer';
-import SearchBar from '../../components/searchBar';
+import SearchBar from '../../containers/searchBar';
 
 test('renders Correctly', () => {
   const Placeholder = () => (
